Updated libraries (airtable, webpack, etc.)  Cleaned up some code.  Converted example to using TS.  Added prettier.
