+++ .prettierrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  printWidth: 120,
+  singleQuote: true,
+  trailingComma: 'none',
+  semi: true,
+  arrowParens: 'always'
+};
