+++ example/src/@types/airtable.d.ts
@@ -1,0 +1,4 @@
+declare module '*.airtable' {
+  const content: any;
+  export default content;
+}
