+++ package.json
@@ -1,21 +1,27 @@
 {
   "name": "airtable-loader",
-  "version": "0.3.1",
-  "description": "",
-  "main": "index.js",
+  "version": "1.0.0",
+  "description": "Webpack loader for injecting Airtable data",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 0"
+    "fix": "prettier \"{src,lib}/**/*.{js,jsx,ts,tsx,md,json}\" \"*.{js,json,md}\" --write"
   },
-  "keywords": ["airtable", "webpack"],
+  "keywords": [
+    "airtable",
+    "webpack"
+  ],
   "author": "Eric Erickson",
   "repository": "https://github.com/steamedcotton/airtable-loader",
   "license": "MIT",
   "dependencies": {
-    "airtable": "^0.8.0",
+    "airtable": "^0.11.1",
     "bluebird": "^3.7.2",
-    "chalk": "^3.0.0",
-    "filesize": "^6.0.1",
-    "loader-utils": "^1.2.3",
-    "schema-utils": "^2.6.1"
+    "chalk": "^4.1.2",
+    "filesize": "^8.0.6",
+    "lodash": "^4.17.21",
+    "schema-utils": "^4.0.0"
+  },
+  "devDependencies": {
+    "prettier": "^2.5.1"
   }
 }
